Move FieldCommunicator<T> to it's own header file.
