+++ src/wdmcpl/client.h
@@ -1,0 +1,4 @@
+#ifndef WDM_COUPLING_CLIENT_H
+#define WDM_COUPLING_CLIENT_H
+
+#endif //WDM_COUPLING_CLIENT_H
