+++ src/wdmcpl/server.h
@@ -1,0 +1,4 @@
+#ifndef WDM_COUPLING_SERVER_H
+#define WDM_COUPLING_SERVER_H
+
+#endif //WDM_COUPLING_SERVER_H
